requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
Pillow
